requirements.txt
fastapi~=0.109.2
uvicorn[standard]~=0.27.0.post1
transformers~=4.37.2
pillow~=10.2.0
torch~=2.2.0
h11~=0.14.0
pip~=22.3.1
wheel~=0.38.4
Jinja2~=3.1.3
requests~=2.31.0
colorama~=0.4.6
PyYAML~=6.0.1
sniffio~=1.3.0
numpy~=1.26.4
setuptools~=65.5.1
regex~=2023.12.25
click~=8.1.7
fsspec~=2024.2.0
certifi~=2024.2.2
starlette~=0.36.3
pydantic~=2.6.1
websockets~=12.0
anyio~=4.2.0
urllib3~=2.2.0
httptools~=0.6.1
watchfiles~=0.21.0
idna~=3.6
packaging~=23.2
tokenizers~=0.15.1
safetensors~=0.4.2
filelock~=3.13.1
tqdm~=4.66.1
torchvision~=0.17.0
torchaudio~=2.2.0